+++ CHANGELOG.md
@@ -86,4 +86,5 @@
 - Removed `multi` parameter from RocketRequest method
 - Added end2end test on example
 - Renamed `multi` `RocketModel` field to `all`
+- Added `targetData` parameter on `request` method of `RocketRequest`
 - Added `forEach` method to `Rocket`